// src/app/employee/model/employee.model.ts
export interface Employee {
    city?: string | null;
    email?: string | null;
    mgrId?: number | null;
    notes?: string | null;
    phone?: string | null;
    photo?: string | null;
    title?: string | null;
    mobile?: string | null;
    region?: string | null,
    address?: string | null,
    country?: string | null,
    entityId: number ,
    hireDate?: string | Date | null,
    lastname?: string,
    birthDate?: string | Date | null,
    extension?: string | null,
    firstname?: string,
    photoPath?: string | null,
    postalCode?: string | null,
    titleOfCourtesy?: string | null
} 